# APIGateway/API-Docs/openapi_v1.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "NURUPE <br>Realizado por:<br><ul><li><p>Carlos Garcia</p></li>José Alves<li><p>",
    "version": "1.3.0",
    "title": "Gerir parque de veiculos"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
...
  "paths": {
...
    "/parque": {
      "get": {
        "tags": ["Parques"],
        "summary": "Obter todas os parques",
        "description": "",
        "operationId": "obter_parques",
        "responses": {
          "200": {
            "description": "Lista de lugares",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/objParqueResponse200"
                }
              }
            }
          },
          "404": {
            "description": "Erro na pesquisa"
          }
        }
      },
      "post": {
        "tags": ["Parques"],
        "summary": "Criar um parque",
        "description": "",
        "operationId": "criar_parque",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/objParqueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parque criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/objRegistarResponse200"
                }
              }
            }
          },
          "400": {
            "description": "Não foram preenchidos todos os campos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/objLoginResponse400"
                }
              }
            }
          },
          "401": {
            "description": "Não tem autorização para criar o parque",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/objLoginResponse400"
                }
              }
            }
          },
          "404": {
            "description": "Erro ao criar o parque"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "objParqueRequest": {
        "$ref": "#/components/schemas/objParque"
      },
...
      "objLoginResponse400": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "objRegistarResponse200": {
        "type": "object",
        "properties": {
          "tipo": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "objParqueResponse200": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/objParque"
        }
      },
...
      "objParque": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "rua": {
            "type": "string"
          },
          "lotacao": {
            "type": "number"
          },
          "estado": {
            "type": "number"
          }
        }
      },
...
    }
  }
}